+++ apps/connect/release_notes.md
@@ -1,5 +1,10 @@
 # ftrack Connect release Notes
 
+## Upcoming
+
+* [changed] Re-structure utils and move general ones to ftrack-utils library.
+* [changed] config.json file renamed to credentials.json
+
 ## v3.0.0
 2024-02-16